refactor: убраны избыточные дублирующие методы репозиториев с без Pageable в аргументах, исправлены все связанные методы в контроллерах и сервисах. Объект Pageable pageable создается в слое контроллеров и передается в сервисный слой; fix: название переменных в аргументах метода findByNameOrDescriptionContainsIgnoreCaseAndAvailable() исправлены refactor: Для случаев, когда было обращение к базе методом findById() для проверки наличия сущности в БД, но сам результаты в дальнейшем коде не использовался, в сервисах добавлены приватные методы проверки, в которых вызывается метод репозитория existById()

+++ src/main/java/ru/practicum/shareit/user/services/UserServiceImpl.java
@@ -33,7 +33,7 @@ public class UserServiceImpl implements UserService {
     public UserDto findById(Long id) throws EntityNotFoundException {
         User user = userRepository.findById(id).orElseThrow(
                 () -> new EntityNotFoundException(String.format("%s with id = %d does not exist in database", "User", id)));
-        log.debug(String.format("User with id = %d founded", user.getId()));
+        log.debug(String.format("User with id = %d founded", id));
         return userMapper.modelToDto(user);
     }
 
@@ -62,9 +62,12 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteById(Long id) throws EntityNotFoundException {
-        findById(id);
-        userRepository.deleteById(id);
-        log.debug(String.format("User with id = %d deleted", id));
+        if (userRepository.existsById(id)) {
+            userRepository.deleteById(id);
+            log.debug(String.format("User with id = %d deleted", id));
+        } else {
+            throw new EntityNotFoundException(String.format("%s with id = %d does not exist in database", "User", id));
+        }
     }
 
     @Override
